Fix stroke bounds for paths
The check was being done backwards-- we should have been checking if
values were less than the current minimum with padding subtracted from
the *values*, and greater than the current maximum with padding added,
but we were subtracting padding from the *minimum* (and adding padding
to the maximum). Instead, we now add padding to the minimum and subtract
it from the maximum, which is equivalent to subtracting padding from the
values when comparing them to the minimum and adding padding to the
values when comparing them to the maximum.